int to long

+++ calculus/pi_number.c
@@ -13,7 +13,7 @@
 double pi1() {
     long k = 0, n = LOOP_N;
     double i, j;
-    for (int m = 0; m < n; m++) {
+    for (long m = 0; m < n; m++) {
         i = (double)rand() / RAND_MAX;
         j = (double)rand() / RAND_MAX;
         if (i * i + j * j < 1.) k++;
@@ -26,7 +26,7 @@ double pi1() {
 double pi2() {
     double pi_qv = 1.0;
     int flag = -1;
-    for (int i = 1; i < LOOP_N; i++) {
+    for (long i = 1; i < LOOP_N; i++) {
         pi_qv += flag * (1./(2 * i + 1));
         flag = -flag;
     }
